fixed pipelines, added additional tests
