+++ tests/integration/Storefront/Theme/ThemeTest.php
@@ -29,6 +29,7 @@ use Shopware\Storefront\Theme\ThemeCollection;
 use Shopware\Storefront\Theme\ThemeCompiler;
 use Shopware\Storefront\Theme\ThemeEntity;
 use Shopware\Storefront\Theme\ThemeLifecycleService;
+use Shopware\Storefront\Theme\ThemeMergedConfigBuilder;
 use Shopware\Storefront\Theme\ThemeService;
 use Shopware\Tests\Integration\Storefront\Theme\fixtures\SimpleTheme\SimpleTheme;
 use Shopware\Tests\Integration\Storefront\Theme\fixtures\SimpleThemeConfigInheritance\SimpleThemeConfigInheritance;
@@ -674,7 +675,8 @@ class ThemeTest extends TestCase
             static::getContainer()->get(Connection::class),
             static::getContainer()->get(SystemConfigService::class),
             static::getContainer()->get('messenger.default_bus'),
-            static::getContainer()->get(NotificationService::class)
+            static::getContainer()->get(NotificationService::class),
+            static::getContainer()->get(ThemeMergedConfigBuilder::class),
         );
         $themeService->updateTheme(
             $childTheme->getId(),
